hybrid_pcg_project/requirements.txt
# Core dependencies for Hybrid PCG Classification Project (Python 3.13 compatible)
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
wfdb>=4.1.0

# Deep learning (CPU versions for compatibility)
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0

# Signal processing and feature extraction
PyWavelets>=1.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data handling
h5py>=3.9.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional: Metrics and evaluation
# tensorboard>=2.13.0
# wandb>=0.15.0
